% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHT11.R
\docType{class}
\name{DHT11-class}
\alias{DHT11-class}
\title{DHT11 Class Representation.}
\description{
DHT11 Class Representation.
}
\section{Slots}{

\describe{
\item{\code{pin}}{The DHT11's assigned pin.}

\item{\code{dht11}}{The internal representation of the DHT11 sensor.}
}}

