% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dht11_sensor.R
\name{get_dht11_sensor}
\alias{get_dht11_sensor}
\title{Get a DHT11 sensor instance.}
\usage{
get_dht11_sensor(pin)
}
\arguments{
\item{pin}{An integer of length one with the GPIO pin that the sensor is
connected to.}
}
\value{
A DHT11 S4 object, this object will be set to work with the selected
  sensor.
}
\description{
Returns a DHT11 sensor instance ready to be read.
}
\examples{
\dontrun{
# Get a DHT11 instance connected at GPIO 14.
dht11_14 <- get_dht11_sensor(14)
# Get a DHT11 instance connected at GPIO 18.
dht11_18 <- get_dht11_sensor(18)
}

}
