% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_humidity.R
\name{get_humidity}
\alias{get_humidity}
\alias{get_humidity,DHT11-method}
\title{Get the humidity of a sensor instance.}
\usage{
get_humidity(x, ...)

\S4method{get_humidity}{DHT11}(x, max_retries = 10, ...)
}
\arguments{
\item{x}{A sensor instance.}

\item{...}{further arguments passed to or from other methods.}

\item{max_retries}{A numeric of length one indicating how many times it
should retry in case of issues.}
}
\value{
A numeric of length one with the read humidity. In case of error, it
  will return `NA`.
}
\description{
Reads the humidity of a given sensor instance.
}
\section{Methods (by class)}{
\itemize{
\item \code{DHT11}: 
}}

\examples{
\dontrun{
# Get a DHT11 sensor instance connected at GPIO 14.
sensor <- get_dht11_sensor(14)
# Read the humidity.
get_humidity(sensor)
}

}
