% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_temperature.R
\name{get_temperature}
\alias{get_temperature}
\alias{get_temperature,DHT11-method}
\title{Get the temperature of a sensor instance.}
\usage{
get_temperature(x, ...)

\S4method{get_temperature}{DHT11}(x, scale = "celsius", max_retries = 10, ...)
}
\arguments{
\item{x}{A sensor instance.}

\item{...}{further arguments passed to or from other methods.}

\item{scale}{One of `"celsius"`, `"fahrenheit"` or `"kelvin"`, indicating
the desired return scale.}

\item{max_retries}{A numeric of length one indicating how many times it
should retry in case of issues.}
}
\value{
A numeric of length one with the read temperature in the selected
  `scale`. In case of error, it will return `NA`.
}
\description{
Reads the temperature of a given sensor instance.
}
\section{Methods (by class)}{
\itemize{
\item \code{DHT11}: 
}}

\examples{
\dontrun{
# Get a DHT11 sensor instance connected at GPIO 14.
sensor <- get_dht11_sensor(14)
# Read the temperature in Celsius and in Fahrenheit.
get_temperature(sensor)
get_temperature(sensor, "fahrenheit")
}

}
