% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sensor.R
\name{read_sensor}
\alias{read_sensor}
\alias{read_sensor,DHT11-method}
\title{Read the data of a sensor instance.}
\usage{
read_sensor(x, ...)

\S4method{read_sensor}{DHT11}(x, max_retries = 10, ...)
}
\arguments{
\item{x}{A sensor instance.}

\item{...}{further arguments passed to or from other methods.}

\item{max_retries}{A numeric of length one indicating how many times it
should retry in case of issues.}
}
\description{
Reads the data of a given sensor instance.
}
\section{Methods (by class)}{
\itemize{
\item \code{DHT11}: 
}}

\examples{
\dontrun{
# Get a DHT11 sensor instance connected at GPIO 14.
dht11 <- get_dht11_sensor(14)
# Read the sensor's data.
read_sensor(dht11)
}

}
